test(rust-autofixer): clarify pinocchio test fixtures
